JSAO9: extracted functions in allocationInfo, added spinners on loading, removed unnecessary if statement
