+++ frontend-opal/src/components/PortfolioContainer/ValueInfo/ValueInfo.js
@@ -2,17 +2,14 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import { View, Text } from 'react-native';
-import { Card, CardItem, Button } from 'native-base';
+import { Card, CardItem, Button, Spinner } from 'native-base';
 import getTotalInvestment from 'components/common/getTotalInvestment';
 import styles from './styles';
 
 const ValueInfo = ({ portfolio }) => {
-  let totalInvestment = '';
-  if (portfolio[0]) {
-    totalInvestment = getTotalInvestment(portfolio)
-      .toString()
-      .replace(/\B(?=(\d{3})+(?!\d))/g, ',');
-  }
+  const totalInvestment = getTotalInvestment(portfolio)
+    .toString()
+    .replace(/\B(?=(\d{3})+(?!\d))/g, ',');
 
   return (
     <CardItem>
@@ -22,7 +19,11 @@ const ValueInfo = ({ portfolio }) => {
         <Card>
           <View style={styles.investment}>
             <Text style={styles.name}>Total Investment</Text>
-            <Text style={styles.amount}>{`$ ${totalInvestment}`}</Text>
+            {totalInvestment === '0' ? (
+              <Spinner color="#e17055" />
+            ) : (
+              <Text style={styles.amount}>{`$ ${totalInvestment}`}</Text>
+            )}
           </View>
 
           <View style={styles.valueButton}>
