+++ frontend-opal/src/components/common/getTotalInvestment.js
@@ -1,7 +1,7 @@
 function getTotalInvestment(fetchPortfolio) {
   const totalInvestment = fetchPortfolio
     .map(instrument => instrument.amount * instrument.marketValue)
-    .reduce((a, b) => a + b);
+    .reduce((a, b) => a + b, 0);
   return totalInvestment;
 }
